Create request package, move api calls and ws, update logging
