+++ .circleci/config.yml
@@ -1,18 +1,16 @@
 # Javascript Node CircleCI 2.0 configuration file
-#
-# Check https://circleci.com/docs/2.0/language-javascript/ for more details
-#
+
 version: 2
 jobs:
   build:
     docker:
       # specify the version you desire here
-      - image: circleci/node:7.10
+      - image: circleci/node:10
 
     working_directory: ~/repo
 
     steps:
       - checkout
 
-      - run: npm i
+      - run: npm ci
       - run: npm test
