+++ package.json
@@ -18,12 +18,12 @@
   "dependencies": {
     "@streammedev/parse-message": "^1.0.1",
     "commander": "^2.16.0",
-    "happiness": "^10.0.2",
     "moment": "^2.22.2",
     "moment-duration-format": "^2.2.2",
-    "npm": "^5.8.0",
     "pino": "^4.17.3",
-    "request": "^2.85.0",
     "ws": "^5.1.1"
+  },
+  "devDependencies": {
+    "happiness": "^10.0.2"
   }
 }
